refactor(glens): split chat.py router by endpoint domain (#1459)
Pure code motion — no behavior change. Same 10 endpoints, same auth,
same handlers. Same mechanical approach as PR #1458 (lens.py split).

Before: chat.py = 1251 LoC / 29 functions / 7 classes / 10 endpoints.
After:  chat.py = 817 LoC focused on the streaming endpoint + its
        private helpers.

Split-out files (all under apps/api/app/modules/glens/routers/):

- sessions.py (112 LoC) — list / get / rename / delete session
- feedback.py (98 LoC)  — POST /chat/feedback thumbs verdict
- opener.py (52 LoC)    — GET /opener chip suggestions
- policy.py (119 LoC)   — POST /policy/apply (create/patch/delete rules)
- config.py (116 LoC)   — POST /guard_config/apply + /spend_config/apply
- _helpers.py (45 LoC)  — _parse_workspace_id, _parse_session_id,
                          _get_session, SessionOut model

Each new file gets its own APIRouter(prefix="/glens", tags=["glens"]).
main.py grows five imports + five include_router calls.

Kept in chat.py:
- glens_chat_stream — the SSE streaming endpoint
- TOOLS derivation from default_registry (#1281)
- _resolve_tools, _stream_synthesis, _guarded_openai_completion
- _build_drilldown, _extract_confirm_envelope, _has_data
- _build_llm_messages, _bg_save_session, _should_refresh_summary
- _llm_config, _load_system_prompt, _SYSTEM
- ChatRequest model, _APOLOGY_PHRASES, _answer_is_apology

Test import updated: tests/glens/test_chat_feedback_endpoint.py now
imports FeedbackIn + submit_feedback from routers.feedback instead of
routers.chat. Other tests (extract_confirm_envelope, tools_catalog)
unchanged — those symbols stayed in chat.py.

Verified:
- All 6 modules import cleanly
- 10 endpoints registered with identical paths + methods
- 70 tools still derive from default_registry
- 18 glens tests pass (feedback, envelope, catalog)
- main.py loads with 81 total routes

Closes #1459.
