+++ apps/api/app/modules/glens/routers/_helpers.py
@@ -1,0 +1,45 @@
+"""Shared helpers + models for the split glens router files.
+
+Extracted from chat.py in #1459. Kept private (leading underscore) —
+consumers are the sibling router modules, not external code.
+"""
+from __future__ import annotations
+
+import uuid
+
+from fastapi import HTTPException
+from pydantic import BaseModel
+from sqlalchemy.orm import Session
+
+from app.modules.glens.models import GlensChatSession
+
+
+class SessionOut(BaseModel):
+    id: str
+    title: str
+    created_at: str
+    has_dashboard: bool
+
+
+def _parse_workspace_id(workspace_id: str) -> uuid.UUID:
+    try:
+        return uuid.UUID(workspace_id)
+    except ValueError:
+        raise HTTPException(status_code=400, detail="Invalid workspace_id")
+
+
+def _parse_session_id(session_id: str) -> uuid.UUID:
+    try:
+        return uuid.UUID(session_id)
+    except ValueError:
+        raise HTTPException(status_code=400, detail="Invalid session_id")
+
+
+def _get_session(db: Session, session_id: str, ws_uuid: uuid.UUID) -> GlensChatSession:
+    session = db.query(GlensChatSession).filter(
+        GlensChatSession.id == _parse_session_id(session_id),
+        GlensChatSession.workspace_id == ws_uuid,
+    ).first()
+    if not session:
+        raise HTTPException(status_code=404, detail="Session not found")
+    return session
