+++ apps/api/app/modules/glens/routers/chat.py
@@ -1,7 +1,9 @@
 """
 POST /glens/chat/stream  — GLens governance assistant (tool-use + prose)
-GET  /glens/sessions     — list sessions
-GET  /glens/sessions/{id} — restore session
+
+Sessions / feedback / opener / policy / config / guard_config / spend_config
+live in sibling router modules (#1459 split). Helpers used across those
+files sit in `_helpers.py`.
 """
 import asyncio
 import json
@@ -9,17 +11,18 @@ import uuid
 from datetime import datetime, timezone
 
 import structlog
-from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException
+from fastapi import APIRouter, BackgroundTasks, Depends
 from fastapi.responses import StreamingResponse
 from pydantic import BaseModel
 from sqlalchemy.orm import Session
 
-from app.core.auth import get_user_id, get_workspace_id, require_permission
+from app.core.auth import get_workspace_id, require_permission
 from app.core.database import get_db
 from app.modules.glens.executor import Executor
-from app.modules.glens.models import GlensChatFeedback, GlensChatSession
-from app.modules.guard.models import GuardConfig, GuardSpendBudget, WorkspaceCustomRule
+from app.modules.glens.models import GlensChatSession
 from app.tools import registrations as _tool_registrations  # noqa: F401  # side-effect: populate default_registry before TOOLS derives
+
+from ._helpers import _get_session, _parse_workspace_id
 
 log = structlog.get_logger(__name__)
 router = APIRouter(prefix="/glens", tags=["glens"])
@@ -621,31 +624,7 @@ def _stream_synthesis(msgs: list[dict], system: str, executor: Executor, on_toke
     return text or "Could not complete the analysis."
 
 
-# ── Session helpers ───────────────────────────────────────────────────────────
-
-def _parse_workspace_id(workspace_id: str) -> uuid.UUID:
-    try:
-        return uuid.UUID(workspace_id)
-    except ValueError:
-        raise HTTPException(status_code=400, detail="Invalid workspace_id")
-
-
-def _parse_session_id(session_id: str) -> uuid.UUID:
-    try:
-        return uuid.UUID(session_id)
-    except ValueError:
-        raise HTTPException(status_code=400, detail="Invalid session_id")
-
-
-def _get_session(db: Session, session_id: str, ws_uuid: uuid.UUID) -> GlensChatSession:
-    session = db.query(GlensChatSession).filter(
-        GlensChatSession.id == _parse_session_id(session_id),
-        GlensChatSession.workspace_id == ws_uuid,
-    ).first()
-    if not session:
-        raise HTTPException(status_code=404, detail="Session not found")
-    return session
-
+# ── Session helpers used only by the stream endpoint ─────────────────────────
 
 def _should_refresh_summary(messages: list[dict]) -> bool:
     user_count = sum(1 for m in messages if m.get("role") == "user")
@@ -692,13 +671,6 @@ class ChatRequest(BaseModel):
     message: str
     session_id: str | None = None
     page_context: str | None = None
-
-
-class SessionOut(BaseModel):
-    id: str
-    title: str
-    created_at: str
-    has_dashboard: bool
 
 
 @router.post("/chat/stream")
@@ -843,409 +815,3 @@ async def glens_chat_stream(
             task.cancel()
 
     return StreamingResponse(generate(), media_type="text/event-stream")
-
-
-# ── Session CRUD ──────────────────────────────────────────────────────────────
-
-@router.get("/sessions", response_model=list[SessionOut])
-def list_sessions(
-    _: str = Depends(require_permission("guard.activity.view_own")),
-    workspace_id: str = Depends(get_workspace_id),
-    db: Session = Depends(get_db),
-):
-    ws_uuid = _parse_workspace_id(workspace_id)
-    sessions = (
-        db.query(GlensChatSession)
-        .filter(GlensChatSession.workspace_id == ws_uuid)
-        .order_by(GlensChatSession.updated_at.desc())
-        .limit(50)
-        .all()
-    )
-    return [
-        SessionOut(id=str(s.id), title=s.title, created_at=s.created_at.isoformat(), has_dashboard=False)
-        for s in sessions
-    ]
-
-
-@router.get("/sessions/{session_id}")
-def get_session(
-    session_id: str,
-    _: str = Depends(require_permission("guard.activity.view_own")),
-    workspace_id: str = Depends(get_workspace_id),
-    db: Session = Depends(get_db),
-):
-    ws_uuid = _parse_workspace_id(workspace_id)
-    session = _get_session(db, session_id, ws_uuid)
-    messages = json.loads(session.messages)
-    user_messages = [
-        {"role": m["role"], "content": m["content"]}
-        for m in messages if m["role"] != "system"
-    ]
-    return {
-        "id": str(session.id),
-        "title": session.title,
-        "messages": user_messages,
-        "spec": None,
-        "created_at": session.created_at.isoformat(),
-    }
-
-
-class SessionTitleUpdate(BaseModel):
-    title: str
-
-
-@router.patch("/sessions/{session_id}", response_model=SessionOut)
-def rename_session(
-    session_id: str,
-    body: SessionTitleUpdate,
-    _: str = Depends(require_permission("guard.activity.view_own")),
-    workspace_id: str = Depends(get_workspace_id),
-    db: Session = Depends(get_db),
-):
-    ws_uuid = _parse_workspace_id(workspace_id)
-    session = _get_session(db, session_id, ws_uuid)
-    title = body.title.strip()
-    if not title:
-        raise HTTPException(status_code=400, detail="Title cannot be empty")
-    session.title = title[:120]
-    session.updated_at = datetime.now(timezone.utc)
-    db.commit()
-    db.refresh(session)
-    return SessionOut(
-        id=str(session.id),
-        title=session.title,
-        created_at=session.created_at.isoformat(),
-        has_dashboard=False,
-    )
-
-
-@router.delete("/sessions/{session_id}", status_code=204)
-def delete_session(
-    session_id: str,
-    _: str = Depends(require_permission("guard.activity.view_own")),
-    workspace_id: str = Depends(get_workspace_id),
-    db: Session = Depends(get_db),
-):
-    ws_uuid = _parse_workspace_id(workspace_id)
-    db.query(GlensChatSession).filter(
-        GlensChatSession.id == _parse_session_id(session_id),
-        GlensChatSession.workspace_id == ws_uuid,
-    ).delete(synchronize_session=False)
-    db.commit()
-
-
-@router.get("/opener")
-def glens_opener(
-    _: str = Depends(require_permission("guard.activity.view_own")),
-    workspace_id: str = Depends(get_workspace_id),
-    db: Session = Depends(get_db),
-):
-    executor = Executor(db, workspace_id)
-    try:
-        kpis = executor._tool_get_governance_kpis()
-    except Exception:
-        kpis = {}
-
-    blocked = kpis.get("blocked_today", 0)
-    events = kpis.get("events_today", 0)
-    blocks_mtd = kpis.get("blocks_mtd", 0)
-    devs = kpis.get("active_developers_today", 0)
-
-    chips: list[str] = []
-    if blocked > 0:
-        chips.append(f"Who was blocked today? ({blocked} block{'s' if blocked != 1 else ''})")
-    else:
-        chips.append("Show me today's Guard activity")
-    if events > 0:
-        chips.append(f"Show the most recent events today ({events} total)")
-    else:
-        chips.append("Show recent Guard events")
-    if blocks_mtd > 0:
-        chips.append(f"How many blocks this month? ({blocks_mtd} so far)")
-    else:
-        chips.append("Cost by AI tool this month")
-    if devs > 1:
-        chips.append(f"Who are the {devs} active developers today?")
-    else:
-        chips.append("Which rule triggered most this week?")
-
-    return {"chips": chips}
-
-
-# ── Policy / config / spend apply endpoints ───────────────────────────────────
-
-class PolicyApplyRequest(BaseModel):
-    action: str
-    draft: dict
-    target_rule_id: str | None = None
-
-
-@router.post("/policy/apply", status_code=201)
-def policy_apply(
-    req: PolicyApplyRequest,
-    _: str = Depends(require_permission("guard.policies.edit")),
-    workspace_id: str = Depends(get_workspace_id),
-    db: Session = Depends(get_db),
-):
-    ws_uuid = _parse_workspace_id(workspace_id)
-
-    if req.action == "create":
-        rule_id = req.draft.get("rule_id")
-        if not rule_id:
-            raise HTTPException(status_code=400, detail="draft.rule_id is required")
-        existing = db.query(WorkspaceCustomRule).filter(
-            WorkspaceCustomRule.workspace_id == ws_uuid,
-            WorkspaceCustomRule.rule_id == rule_id,
-        ).first()
-        if existing:
-            raise HTTPException(status_code=409, detail=f"Rule '{rule_id}' already exists")
-        if pattern := req.draft.get("match_pattern"):
-            import re
-            try:
-                re.compile(pattern)
-            except re.error as e:
-                raise HTTPException(status_code=400, detail=f"Invalid match_pattern regex: {e}")
-        body = {k: v for k, v in req.draft.items() if k not in ("rule_id", "persona")}
-        body["id"] = rule_id
-        rule = WorkspaceCustomRule(
-            workspace_id=ws_uuid,
-            rule_id=rule_id,
-            persona=req.draft.get("persona", "agent"),
-            body=body,
-            enabled=True,
-        )
-        db.add(rule)
-        db.commit()
-        return {"ok": True, "rule_id": rule_id, "action": "created"}
-
-    elif req.action == "patch":
-        from app.modules.guard.routers.policies import _upsert_override
-        rule_id = req.target_rule_id
-        if not rule_id:
-            raise HTTPException(status_code=400, detail="target_rule_id is required for patch")
-        if pattern := req.draft.get("match_pattern"):
-            import re
-            try:
-                re.compile(pattern)
-            except re.error as e:
-                raise HTTPException(status_code=400, detail=f"Invalid match_pattern regex: {e}")
-        custom = db.query(WorkspaceCustomRule).filter(
-            WorkspaceCustomRule.workspace_id == ws_uuid,
-            WorkspaceCustomRule.rule_id == rule_id,
-        ).first()
-        if custom:
-            if "enabled" in req.draft:
-                custom.enabled = req.draft["enabled"]
-            body_patch = {k: v for k, v in req.draft.items() if k != "enabled"}
-            if body_patch:
-                custom.body = {**custom.body, **body_patch}
-            custom.updated_at = datetime.now(timezone.utc)
-            db.commit()
-            return {"ok": True, "rule_id": rule_id, "action": "patched"}
-        touched = False
-        if "enabled" in req.draft:
-            _upsert_override(db, ws_uuid, rule_id, disabled=not req.draft["enabled"])
-            touched = True
-        action_val = req.draft.get("action")
-        msg_val = req.draft.get("message")
-        pat_val = req.draft.get("match_pattern")
-        if action_val or msg_val or pat_val:
-            _upsert_override(db, ws_uuid, rule_id, action=action_val, message=msg_val, match_pattern=pat_val)
-            touched = True
-        if not touched:
-            raise HTTPException(status_code=404, detail=f"Rule '{rule_id}' not found")
-        db.commit()
-        return {"ok": True, "rule_id": rule_id, "action": "patched"}
-
-    elif req.action == "delete":
-        rule_id = req.target_rule_id or req.draft.get("rule_id")
-        if not rule_id:
-            raise HTTPException(status_code=400, detail="target_rule_id is required for delete")
-        rule = db.query(WorkspaceCustomRule).filter(
-            WorkspaceCustomRule.workspace_id == ws_uuid,
-            WorkspaceCustomRule.rule_id == rule_id,
-        ).first()
-        if not rule:
-            raise HTTPException(status_code=404, detail=f"Rule '{rule_id}' not found")
-        db.delete(rule)
-        db.commit()
-        return {"ok": True, "rule_id": rule_id, "action": "deleted"}
-
-    raise HTTPException(status_code=400, detail=f"Unknown action '{req.action}'")
-
-
-class GuardConfigApplyRequest(BaseModel):
-    draft: dict
-
-
-_VALID_ENFORCEMENT = {"active", "warn", "advisory", "off"}
-_VALID_FAIL_MODE = {"fail_open", "fail_closed"}
-
-
-@router.post("/guard_config/apply", status_code=200)
-def guard_config_apply(
-    req: GuardConfigApplyRequest,
-    _: str = Depends(require_permission("guard.settings.edit")),
-    workspace_id: str = Depends(get_workspace_id),
-    db: Session = Depends(get_db),
-):
-    ws_uuid = _parse_workspace_id(workspace_id)
-    cfg = db.query(GuardConfig).filter(GuardConfig.workspace_id == ws_uuid).first()
-    if not cfg:
-        raise HTTPException(status_code=404, detail="Guard not configured for this workspace")
-    if "enforcement_mode" in req.draft:
-        val = req.draft["enforcement_mode"]
-        if val not in _VALID_ENFORCEMENT:
-            raise HTTPException(status_code=400, detail=f"enforcement_mode must be one of: {', '.join(sorted(_VALID_ENFORCEMENT))}")
-        cfg.enforcement_mode = val
-    if "fail_mode" in req.draft:
-        val = req.draft["fail_mode"]
-        if val not in _VALID_FAIL_MODE:
-            raise HTTPException(status_code=400, detail="fail_mode must be 'fail_open' or 'fail_closed'")
-        cfg.fail_mode = val
-    for bool_field in ("advisory_mode", "notify_on_block", "notify_on_budget", "deny_on_error"):
-        if bool_field in req.draft:
-            setattr(cfg, bool_field, bool(req.draft[bool_field]))
-    cfg.updated_at = datetime.now(timezone.utc)
-    db.commit()
-    return {"ok": True, "action": "patched"}
-
-
-class SpendConfigApplyRequest(BaseModel):
-    monthly_limit_usd: float
-    hard_limit_usd: float | None = None
-    alert_threshold_pct: int = 80
-    email: str | None = None
-
-
-@router.post("/spend_config/apply", status_code=201)
-def spend_config_apply(
-    req: SpendConfigApplyRequest,
-    _: str = Depends(require_permission("guard.spend.budgets.edit")),
-    workspace_id: str = Depends(get_workspace_id),
-    db: Session = Depends(get_db),
-):
-    from app.modules.guard.models import GuardSession as _GuardSession
-    ws_uuid = _parse_workspace_id(workspace_id)
-
-    clerk_user_id: str | None = None
-    if req.email:
-        session_row = (
-            db.query(_GuardSession)
-            .filter(
-                _GuardSession.workspace_id == ws_uuid,
-                _GuardSession.user_email == req.email,
-                _GuardSession.clerk_user_id.isnot(None),
-            )
-            .order_by(_GuardSession.started_at.desc())
-            .first()
-        )
-        if not session_row:
-            raise HTTPException(status_code=404, detail=f"No Guard sessions found for {req.email}")
-        clerk_user_id = session_row.clerk_user_id
-
-    existing = db.query(GuardSpendBudget).filter(
-        GuardSpendBudget.workspace_id == ws_uuid,
-        GuardSpendBudget.clerk_user_id == clerk_user_id,
-    ).first()
-    now = datetime.now(timezone.utc)
-    if existing:
-        existing.monthly_limit_usd = req.monthly_limit_usd
-        existing.hard_limit_usd = req.hard_limit_usd
-        existing.alert_threshold_pct = req.alert_threshold_pct
-        existing.updated_at = now
-        db.commit()
-        action = "updated"
-    else:
-        db.add(GuardSpendBudget(
-            workspace_id=ws_uuid,
-            clerk_user_id=clerk_user_id,
-            monthly_limit_usd=req.monthly_limit_usd,
-            hard_limit_usd=req.hard_limit_usd,
-            alert_threshold_pct=req.alert_threshold_pct,
-        ))
-        db.commit()
-        action = "created"
-
-    scope = "workspace" if req.email is None else f"developer:{req.email}"
-    return {"ok": True, "action": action, "scope": scope}
-
-
-# ── Feedback (thumbs up/down per message) ────────────────────────────────────
-
-class FeedbackIn(BaseModel):
-    session_id: str
-    message_id: str                    # position/id of the assistant message
-    verdict: str                        # "up" | "down"
-    comment: str | None = None
-
-
-@router.post("/chat/feedback")
-def submit_feedback(
-    req: FeedbackIn,
-    _: str = Depends(require_permission("guard.activity.view_own")),
-    workspace_id: str = Depends(get_workspace_id),
-    user_id: str | None = Depends(get_user_id),
-    db: Session = Depends(get_db),
-):
-    """Record thumbs up/down for one assistant message. Upsert on
-    (session_id, message_id, clerk_user_id) — a user can flip their vote
-    or add a comment; latest wins.
-
-    Downstream: aggregated feedback feeds the LLM tuning + prompt
-    regression loop (nightly rollup, not part of this endpoint).
-    """
-    if req.verdict not in ("up", "down"):
-        raise HTTPException(status_code=400, detail="verdict must be 'up' or 'down'")
-    if not req.message_id.strip():
-        raise HTTPException(status_code=400, detail="message_id is required")
-    if req.comment is not None and len(req.comment) > 2000:
-        raise HTTPException(status_code=400, detail="comment must be <= 2000 chars")
-
-    ws_uuid = _parse_workspace_id(workspace_id)
-    try:
-        session_uuid = uuid.UUID(req.session_id)
-    except (ValueError, TypeError):
-        raise HTTPException(status_code=400, detail="session_id is not a UUID")
-
-    # Confirm the session belongs to this workspace so users can't leave
-    # feedback on someone else's session by guessing the id.
-    session = (
-        db.query(GlensChatSession)
-        .filter(
-            GlensChatSession.id == session_uuid,
-            GlensChatSession.workspace_id == ws_uuid,
-        )
-        .first()
-    )
-    if session is None:
-        raise HTTPException(status_code=404, detail="session not found")
-
-    existing = (
-        db.query(GlensChatFeedback)
-        .filter(
-            GlensChatFeedback.session_id == session_uuid,
-            GlensChatFeedback.message_id == req.message_id,
-            GlensChatFeedback.clerk_user_id == user_id,
-        )
-        .first()
-    )
-
-    now = datetime.now(timezone.utc)
-    if existing:
-        existing.verdict = req.verdict
-        existing.comment = req.comment
-        existing.updated_at = now
-        db.commit()
-        return {"ok": True, "action": "updated", "verdict": req.verdict}
-
-    db.add(GlensChatFeedback(
-        workspace_id=ws_uuid,
-        session_id=session_uuid,
-        message_id=req.message_id,
-        verdict=req.verdict,
-        comment=req.comment,
-        clerk_user_id=user_id,
-    ))
-    db.commit()
-    return {"ok": True, "action": "created", "verdict": req.verdict}
