+++ apps/api/app/modules/glens/routers/config.py
@@ -1,0 +1,116 @@
+"""Guard config + spend budget apply endpoints.
+
+Split out of chat.py in #1459.
+"""
+from __future__ import annotations
+
+from datetime import datetime, timezone
+
+from fastapi import APIRouter, Depends, HTTPException
+from pydantic import BaseModel
+from sqlalchemy.orm import Session
+
+from app.core.auth import get_workspace_id, require_permission
+from app.core.database import get_db
+from app.modules.guard.models import GuardConfig, GuardSpendBudget
+
+from ._helpers import _parse_workspace_id
+
+router = APIRouter(prefix="/glens", tags=["glens"])
+
+
+class GuardConfigApplyRequest(BaseModel):
+    draft: dict
+
+
+_VALID_ENFORCEMENT = {"active", "warn", "advisory", "off"}
+_VALID_FAIL_MODE = {"fail_open", "fail_closed"}
+
+
+@router.post("/guard_config/apply", status_code=200)
+def guard_config_apply(
+    req: GuardConfigApplyRequest,
+    _: str = Depends(require_permission("guard.settings.edit")),
+    workspace_id: str = Depends(get_workspace_id),
+    db: Session = Depends(get_db),
+):
+    ws_uuid = _parse_workspace_id(workspace_id)
+    cfg = db.query(GuardConfig).filter(GuardConfig.workspace_id == ws_uuid).first()
+    if not cfg:
+        raise HTTPException(status_code=404, detail="Guard not configured for this workspace")
+    if "enforcement_mode" in req.draft:
+        val = req.draft["enforcement_mode"]
+        if val not in _VALID_ENFORCEMENT:
+            raise HTTPException(status_code=400, detail=f"enforcement_mode must be one of: {', '.join(sorted(_VALID_ENFORCEMENT))}")
+        cfg.enforcement_mode = val
+    if "fail_mode" in req.draft:
+        val = req.draft["fail_mode"]
+        if val not in _VALID_FAIL_MODE:
+            raise HTTPException(status_code=400, detail="fail_mode must be 'fail_open' or 'fail_closed'")
+        cfg.fail_mode = val
+    for bool_field in ("advisory_mode", "notify_on_block", "notify_on_budget", "deny_on_error"):
+        if bool_field in req.draft:
+            setattr(cfg, bool_field, bool(req.draft[bool_field]))
+    cfg.updated_at = datetime.now(timezone.utc)
+    db.commit()
+    return {"ok": True, "action": "patched"}
+
+
+class SpendConfigApplyRequest(BaseModel):
+    monthly_limit_usd: float
+    hard_limit_usd: float | None = None
+    alert_threshold_pct: int = 80
+    email: str | None = None
+
+
+@router.post("/spend_config/apply", status_code=201)
+def spend_config_apply(
+    req: SpendConfigApplyRequest,
+    _: str = Depends(require_permission("guard.spend.budgets.edit")),
+    workspace_id: str = Depends(get_workspace_id),
+    db: Session = Depends(get_db),
+):
+    from app.modules.guard.models import GuardSession as _GuardSession
+    ws_uuid = _parse_workspace_id(workspace_id)
+
+    clerk_user_id: str | None = None
+    if req.email:
+        session_row = (
+            db.query(_GuardSession)
+            .filter(
+                _GuardSession.workspace_id == ws_uuid,
+                _GuardSession.user_email == req.email,
+                _GuardSession.clerk_user_id.isnot(None),
+            )
+            .order_by(_GuardSession.started_at.desc())
+            .first()
+        )
+        if not session_row:
+            raise HTTPException(status_code=404, detail=f"No Guard sessions found for {req.email}")
+        clerk_user_id = session_row.clerk_user_id
+
+    existing = db.query(GuardSpendBudget).filter(
+        GuardSpendBudget.workspace_id == ws_uuid,
+        GuardSpendBudget.clerk_user_id == clerk_user_id,
+    ).first()
+    now = datetime.now(timezone.utc)
+    if existing:
+        existing.monthly_limit_usd = req.monthly_limit_usd
+        existing.hard_limit_usd = req.hard_limit_usd
+        existing.alert_threshold_pct = req.alert_threshold_pct
+        existing.updated_at = now
+        db.commit()
+        action = "updated"
+    else:
+        db.add(GuardSpendBudget(
+            workspace_id=ws_uuid,
+            clerk_user_id=clerk_user_id,
+            monthly_limit_usd=req.monthly_limit_usd,
+            hard_limit_usd=req.hard_limit_usd,
+            alert_threshold_pct=req.alert_threshold_pct,
+        ))
+        db.commit()
+        action = "created"
+
+    scope = "workspace" if req.email is None else f"developer:{req.email}"
+    return {"ok": True, "action": action, "scope": scope}
