+++ apps/api/app/modules/glens/routers/sessions.py
@@ -1,0 +1,112 @@
+"""Session CRUD for glens chat — list, get, rename, delete.
+
+Split out of chat.py in #1459.
+"""
+from __future__ import annotations
+
+import json
+from datetime import datetime, timezone
+
+from fastapi import APIRouter, Depends, HTTPException
+from pydantic import BaseModel
+from sqlalchemy.orm import Session
+
+from app.core.auth import get_workspace_id, require_permission
+from app.core.database import get_db
+from app.modules.glens.models import GlensChatSession
+
+from ._helpers import (
+    SessionOut,
+    _get_session,
+    _parse_session_id,
+    _parse_workspace_id,
+)
+
+router = APIRouter(prefix="/glens", tags=["glens"])
+
+
+class SessionTitleUpdate(BaseModel):
+    title: str
+
+
+@router.get("/sessions", response_model=list[SessionOut])
+def list_sessions(
+    _: str = Depends(require_permission("guard.activity.view_own")),
+    workspace_id: str = Depends(get_workspace_id),
+    db: Session = Depends(get_db),
+):
+    ws_uuid = _parse_workspace_id(workspace_id)
+    sessions = (
+        db.query(GlensChatSession)
+        .filter(GlensChatSession.workspace_id == ws_uuid)
+        .order_by(GlensChatSession.updated_at.desc())
+        .limit(50)
+        .all()
+    )
+    return [
+        SessionOut(id=str(s.id), title=s.title, created_at=s.created_at.isoformat(), has_dashboard=False)
+        for s in sessions
+    ]
+
+
+@router.get("/sessions/{session_id}")
+def get_session(
+    session_id: str,
+    _: str = Depends(require_permission("guard.activity.view_own")),
+    workspace_id: str = Depends(get_workspace_id),
+    db: Session = Depends(get_db),
+):
+    ws_uuid = _parse_workspace_id(workspace_id)
+    session = _get_session(db, session_id, ws_uuid)
+    messages = json.loads(session.messages)
+    user_messages = [
+        {"role": m["role"], "content": m["content"]}
+        for m in messages if m["role"] != "system"
+    ]
+    return {
+        "id": str(session.id),
+        "title": session.title,
+        "messages": user_messages,
+        "spec": None,
+        "created_at": session.created_at.isoformat(),
+    }
+
+
+@router.patch("/sessions/{session_id}", response_model=SessionOut)
+def rename_session(
+    session_id: str,
+    body: SessionTitleUpdate,
+    _: str = Depends(require_permission("guard.activity.view_own")),
+    workspace_id: str = Depends(get_workspace_id),
+    db: Session = Depends(get_db),
+):
+    ws_uuid = _parse_workspace_id(workspace_id)
+    session = _get_session(db, session_id, ws_uuid)
+    title = body.title.strip()
+    if not title:
+        raise HTTPException(status_code=400, detail="Title cannot be empty")
+    session.title = title[:120]
+    session.updated_at = datetime.now(timezone.utc)
+    db.commit()
+    db.refresh(session)
+    return SessionOut(
+        id=str(session.id),
+        title=session.title,
+        created_at=session.created_at.isoformat(),
+        has_dashboard=False,
+    )
+
+
+@router.delete("/sessions/{session_id}", status_code=204)
+def delete_session(
+    session_id: str,
+    _: str = Depends(require_permission("guard.activity.view_own")),
+    workspace_id: str = Depends(get_workspace_id),
+    db: Session = Depends(get_db),
+):
+    ws_uuid = _parse_workspace_id(workspace_id)
+    db.query(GlensChatSession).filter(
+        GlensChatSession.id == _parse_session_id(session_id),
+        GlensChatSession.workspace_id == ws_uuid,
+    ).delete(synchronize_session=False)
+    db.commit()
